% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{xml_metadata}
\alias{xml_metadata}
\alias{xml_version}
\alias{xml_encoding}
\alias{xml_standalone}
\title{Document metadata}
\usage{
xml_version(x)

xml_encoding(x)

xml_standalone(x)
}
\arguments{
\item{x}{A \code{zuxml_document}.}
}
\value{
A length-1 vector.
}
\description{
Document metadata
}
\examples{
xml_encoding(xml_parse('<?xml version="1.0" encoding="UTF-8"?><a/>'))
}
